Generalize and clean up Segment structure
I realized that when defining structures of higher length, we would end
up with structures that duplicate information.  Instead, we use a
general form for "path" structures---a Vec.  The advantage of doing
this is that we just need to run a few small computations to get this
working.

Vec's are assumed to be in the order of traversal, so care must be
taken to ensure that points are in order.  (BTreeMap would sort points
by coordinate, and would therefore be incorrect.)

Also, position_at() calculations now return an Option.  Most of the
time this will be Some(value), but in cases where the length of a Path
is strictly zero, this value is None.  This is not an Err condition,
just a confusing one.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>