Add second-derivative Hermite basis functions
As you might expect, this is supposed to satisfy the following
constraint:

  H_x^5''(0, x) == 1 iff x == 2, and
  H_x^5''(1, x) == 1 iff x == 4.

Note that I can't keep going with this---quintic hermite basis
functions only work up to acceleration.  The next derivative, for
example, would have H_x^5''(0, 0) == -60, which is wrong.  To get
higher degrees, I need to develop septic Hermite basis functions.

This works for most motion planning, so I'm calling it here.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>